create language_utils.py
